kageco-299 エンティティのパッケージ構造を再編成 ・エンティティクラスを目的別に`pub`、`household`、`business`といったパッケージに移動し、整理した。 ・一部クラスにおいて、関連する他クラス(`Group`, `User`, など)のインポートを修正した。 ・パッケージ構造の明確化により、クラスの役割が分かりやすくなり、保守性が向上した対応である。
