--- backend/api/src/main/java/kageco/api/entity/DepositCategory.java
@@ -1,37 +1,0 @@
-package kageco.api.entity;
-
-import lombok.Getter;
-import lombok.Setter;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.MapsId;
-import jakarta.persistence.OneToOne;
-import jakarta.persistence.Table;
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Size;
-
-@Getter
-@Setter
-@Entity
-@Table(name = "deposit_category", schema = "household")
-public class DepositCategory {
-    @Id
-    @Size(max = 26)
-    @Column(name = "category_id", nullable = false, length = 26)
-    private String categoryId;
-
-    @MapsId
-    @OneToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(name = "category_id", nullable = false)
-    private Category category;
-
-    @NotNull
-    @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(name = "group_id", nullable = false)
-    private Group group;
-
-}
