+++ backend/api/src/main/java/kageco/api/entity/business/DailyAttendanceLog.java
@@ -1,4 +1,4 @@
-package kageco.api.entity;
+package kageco.api.entity.business;
 
 import java.time.OffsetDateTime;
