+++ backend/api/src/main/java/kageco/api/entity/household/Category.java
@@ -1,4 +1,4 @@
-package kageco.api.entity;
+package kageco.api.entity.household;
 
 import java.util.LinkedHashSet;
 import java.util.Set;
@@ -18,6 +18,7 @@ import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
+import kageco.api.entity.pub.Group;
 
 @Getter
 @Setter
