+++ backend/api/src/main/java/kageco/api/entity/household/ConditionSession.java
@@ -1,4 +1,4 @@
-package kageco.api.entity;
+package kageco.api.entity.household;
 
 import java.util.List;
 import java.util.Map;
@@ -15,6 +15,7 @@ import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import jakarta.validation.constraints.NotNull;
+import kageco.api.entity.pub.Group;
 
 @Getter
 @Setter
