+++ backend/api/src/main/java/kageco/api/entity/household/CreditCardDetail.java
@@ -1,8 +1,10 @@
-package kageco.api.entity;
+package kageco.api.entity.household;
 
 import java.math.BigDecimal;
 import java.time.LocalDate;
 
+import org.hibernate.annotations.OnDelete;
+import org.hibernate.annotations.OnDeleteAction;
 import lombok.Getter;
 import lombok.Setter;
 import jakarta.persistence.Column;
@@ -14,12 +16,14 @@ import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
+import kageco.api.entity.pub.Group;
+import kageco.api.entity.pub.User;
 
 @Getter
 @Setter
 @Entity
-@Table(name = "daily_detail", schema = "household")
-public class DailyDetail {
+@Table(name = "credit_card_detail", schema = "household")
+public class CreditCardDetail {
     @Id
     @Size(max = 26)
     @Column(name = "id", nullable = false, length = 26)
@@ -45,22 +49,23 @@ public class DailyDetail {
     private Category category;
 
     @NotNull
-    @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(name = "account_id", nullable = false)
-    private Account account;
-
-    @NotNull
-    @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(name = "user_id", nullable = false)
-    private User user;
-
-    @NotNull
     @Column(name = "amount", nullable = false, precision = 10)
     private BigDecimal amount;
 
     @Size(max = 64)
     @Column(name = "memo", length = 64)
     private String memo;
+
+    @NotNull
+    @ManyToOne(fetch = FetchType.LAZY, optional = false)
+    @OnDelete(action = OnDeleteAction.CASCADE)
+    @JoinColumn(name = "summary_id", nullable = false)
+    private CreditCardSummary summary;
+
+    @NotNull
+    @ManyToOne(fetch = FetchType.LAZY, optional = false)
+    @JoinColumn(name = "user_id", nullable = false)
+    private User user;
 
     @NotNull
     @ManyToOne(fetch = FetchType.LAZY, optional = false)
