+++ backend/api/src/main/java/kageco/api/entity/household/FavoriteFilter.java
@@ -1,4 +1,7 @@
-package kageco.api.entity;
+package kageco.api.entity.household;
+
+import java.util.LinkedHashSet;
+import java.util.Set;
 
 import lombok.Getter;
 import lombok.Setter;
@@ -8,28 +11,33 @@ import jakarta.persistence.FetchType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
+import kageco.api.entity.pub.Group;
 
 @Getter
 @Setter
 @Entity
-@Table(name = "group_application", schema = "public")
-public class GroupApplication {
+@Table(name = "favorite_filter", schema = "household")
+public class FavoriteFilter {
     @Id
     @Size(max = 26)
     @Column(name = "id", nullable = false, length = 26)
     private String id;
+
+    @Size(max = 64)
+    @NotNull
+    @Column(name = "name", nullable = false, length = 64)
+    private String name;
 
     @NotNull
     @ManyToOne(fetch = FetchType.LAZY, optional = false)
     @JoinColumn(name = "group_id", nullable = false)
     private Group group;
 
-    @NotNull
-    @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(name = "application_id", nullable = false)
-    private Application application;
+    @OneToMany(mappedBy = "filter")
+    private Set<FavoriteFilterArg> favoriteFilterArgs = new LinkedHashSet<>();
 
 }
