+++ backend/api/src/main/java/kageco/api/entity/household/FreeeLinkDetail.java
@@ -1,4 +1,4 @@
-package kageco.api.entity;
+package kageco.api.entity.household;
 
 import java.time.OffsetDateTime;
