+++ backend/api/src/main/java/kageco/api/entity/household/Genre.java
@@ -1,4 +1,4 @@
-package kageco.api.entity;
+package kageco.api.entity.household;
 
 import java.util.LinkedHashSet;
 import java.util.Set;
@@ -10,18 +10,22 @@ import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.Id;
+import jakarta.persistence.Index;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
+import kageco.api.entity.pub.Group;
 
 @Getter
 @Setter
 @Entity
-@Table(name = "account", schema = "household")
-public class Account {
+@Table(name = "genre", schema = "household", indexes = {
+    @Index(name = "genre_iocome_type_idx", columnList = "iocome_type")
+})
+public class Genre {
     @Id
     @Size(max = 26)
     @Column(name = "id", nullable = false, length = 26)
@@ -31,6 +35,16 @@ public class Account {
     @NotNull
     @Column(name = "name", nullable = false, length = 50)
     private String name;
+
+    @Size(max = 16)
+    @NotNull
+    @Column(name = "genre_type", nullable = false, length = 16)
+    private String genreType;
+
+    @Size(max = 8)
+    @NotNull
+    @Column(name = "iocome_type", nullable = false, length = 8)
+    private String iocomeType;
 
     @ColumnDefault("true")
     @Column(name = "valid_flag")
@@ -45,10 +59,13 @@ public class Account {
     @JoinColumn(name = "group_id", nullable = false)
     private Group group;
 
-    @OneToMany(mappedBy = "account")
-    private Set<CreditCardSummary> creditCardSummaries = new LinkedHashSet<>();
+    @OneToMany(mappedBy = "genre")
+    private Set<Category> categories = new LinkedHashSet<>();
 
-    @OneToMany(mappedBy = "account")
+    @OneToMany(mappedBy = "genre")
+    private Set<CreditCardDetail> creditCardDetails = new LinkedHashSet<>();
+
+    @OneToMany(mappedBy = "genre")
     private Set<DailyDetail> dailyDetails = new LinkedHashSet<>();
 
 }
