+++ backend/api/src/main/java/kageco/api/entity/household/ImportFileHistory.java
@@ -1,7 +1,6 @@
-package kageco.api.entity;
+package kageco.api.entity.household;
 
-import java.util.LinkedHashSet;
-import java.util.Set;
+import java.time.Instant;
 
 import lombok.Getter;
 import lombok.Setter;
@@ -11,32 +10,43 @@ import jakarta.persistence.FetchType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
+import kageco.api.entity.pub.Group;
 
 @Getter
 @Setter
 @Entity
-@Table(name = "favorite_filter", schema = "household")
-public class FavoriteFilter {
+@Table(name = "import_file_history", schema = "household")
+public class ImportFileHistory {
     @Id
     @Size(max = 26)
     @Column(name = "id", nullable = false, length = 26)
     private String id;
 
-    @Size(max = 64)
+    @Size(max = 128)
     @NotNull
-    @Column(name = "name", nullable = false, length = 64)
-    private String name;
+    @Column(name = "file_name", nullable = false, length = 128)
+    private String fileName;
+
+    @Size(max = 16)
+    @NotNull
+    @Column(name = "file_type", nullable = false, length = 16)
+    private String fileType;
+
+    @NotNull
+    @Column(name = "import_datetime", nullable = false)
+    private Instant importDatetime;
+
+    @Size(max = 26)
+    @NotNull
+    @Column(name = "import_user_id", nullable = false, length = 26)
+    private String importUserId;
 
     @NotNull
     @ManyToOne(fetch = FetchType.LAZY, optional = false)
     @JoinColumn(name = "group_id", nullable = false)
     private Group group;
-
-    @OneToMany(mappedBy = "filter")
-    private Set<FavoriteFilterArg> favoriteFilterArgs = new LinkedHashSet<>();
 
 }
