+++ backend/api/src/main/java/kageco/api/entity/household/Tag.java
@@ -1,4 +1,4 @@
-package kageco.api.entity;
+package kageco.api.entity.household;
 
 import java.util.LinkedHashSet;
 import java.util.Set;
@@ -18,6 +18,7 @@ import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import jakarta.validation.constraints.NotNull;
+import kageco.api.entity.pub.Group;
 
 @Getter
 @Setter
