+++ backend/api/src/main/java/kageco/api/entity/pub/Affiliation.java
@@ -1,6 +1,4 @@
-package kageco.api.entity;
-
-import java.time.Instant;
+package kageco.api.entity.pub;
 
 import lombok.Getter;
 import lombok.Setter;
@@ -17,35 +15,26 @@ import jakarta.validation.constraints.Size;
 @Getter
 @Setter
 @Entity
-@Table(name = "import_file_history", schema = "household")
-public class ImportFileHistory {
+@Table(name = "affiliation", schema = "public")
+public class Affiliation {
     @Id
     @Size(max = 26)
     @Column(name = "id", nullable = false, length = 26)
     private String id;
 
-    @Size(max = 128)
     @NotNull
-    @Column(name = "file_name", nullable = false, length = 128)
-    private String fileName;
-
-    @Size(max = 16)
-    @NotNull
-    @Column(name = "file_type", nullable = false, length = 16)
-    private String fileType;
-
-    @NotNull
-    @Column(name = "import_datetime", nullable = false)
-    private Instant importDatetime;
-
-    @Size(max = 26)
-    @NotNull
-    @Column(name = "import_user_id", nullable = false, length = 26)
-    private String importUserId;
+    @ManyToOne(fetch = FetchType.LAZY, optional = false)
+    @JoinColumn(name = "user_id", nullable = false)
+    private User user;
 
     @NotNull
     @ManyToOne(fetch = FetchType.LAZY, optional = false)
     @JoinColumn(name = "group_id", nullable = false)
     private Group group;
+
+    @Size(max = 16)
+    @NotNull
+    @Column(name = "group_role", nullable = false, length = 16)
+    private String groupRole;
 
 }
