+++ backend/api/src/main/java/kageco/api/entity/pub/App.java
@@ -1,4 +1,4 @@
-package kageco.api.entity;
+package kageco.api.entity.pub;
 
 import org.hibernate.annotations.OnDelete;
 import org.hibernate.annotations.OnDeleteAction;
